main, weather and dependant files organized / created
