+++ lib/components/DayWeather.js
@@ -1,14 +1,9 @@
 import React from 'react';
 
 const DayWeather = () => {
-//   if (!location) {
-//     return (
-//       <div>Please enter a valid City, STATE</div>
-//     );
-//   }
-//
+
   return(
-    <div>Day's Weather</div>
+    <div>Days Weather</div>
   )
 }
