+++ lib/components/HourWeather.js
@@ -1,14 +1,13 @@
 import React, { Component } from 'react';
 
-export default class TodayWeather extends Component {
-  constructor() {
-    super()
+export default class HourWeather extends Component {
 
-  }
+  const HourWeather = ({weather}) => {
 
-  render() {
-    return(
-      <div>TodaysWeather</div>
-    )
+    render() {
+      return(
+        <div>Today</div>
+      )
+    }
   }
 }
