+++ lib/components/TodayWeather.js
@@ -1,7 +1,8 @@
 import React from 'react';
 import HourWeather from './HourWeather'
+import CurrentWeather from './CurrentWeather'
 
-const TodayWeather = ({ location }) => {
+const TodayWeather = ({ location, weather }) => {
   if (!location) {
     return (
       <div>Please enter a valid City, STATE</div>
@@ -9,7 +10,7 @@ const TodayWeather = ({ location }) => {
   }
 
   return(
-    <HourWeather />
+    <HourWeather weather = {weather} []/>
   )
 };
