+++ lib/components/Weather.js
@@ -1,4 +1,6 @@
 import React, { Component } from 'react';
+
+import Controls from './Controls'
 import TodayWeather from './TodayWeather';
 import Forecast from './Forecast';
 
@@ -6,37 +8,16 @@ export default class Weather extends Component {
   constructor() {
     super()
     this.state = {
-      input: '',
-      location: [],
+      weather: []
     }
-  }
-
-  handleClick() {
-    const userLocation = { location: this.state.input, id: Date.now() };
-
-    this.state.location.push(userLocation);
-    this.setState({
-      input: '',
-      location: this.state.location,
-    });
   }
 
   render() {
     return(
       <div>
-        <input type='text'
-               placeholder='Enter City, STATE' value={this.state.input}
-               onChange= { (event) => { this.setState({ input: event.target.value }); }}
-        />
-
-        <input type='submit'
-                 onClick={this.handleClick.bind(this)}
-          />
-
-          <TodayWeather location = { this.state.location }
-          />
-
-          <Forecast />
+        <Controls weather     = {this.state.weather} />
+        <TodayWeather weather = {this.state.weather} />
+        <Forecast weather     = {this.state.weather} />
       </div>
     )
   }
